# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agro Vision Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Crop {
  id: string;
  name: string;
  scientificName: string;
  description: string;
  imageUrl: string;
  commonDiseases: string[];
  season: string;
  difficulty: string;
}

export interface Scan {
  id: number;
  cropId: string;
  cropName: string;
  imageDataUrl: string;
  /** healthy | diseased | invalid */
  status: string;
  diseaseName?: string | null;
  confidence?: number | null;
  severity?: string | null;
  summary: string;
  symptoms: string[];
  treatment: string[];
  prevention: string[];
  createdAt: string;
}

export interface ScanError {
  error: string;
  reason: string;
}

export type DashboardSummaryWeeklyTrendItem = {
  day: string;
  healthy: number;
  diseased: number;
};

export type DashboardSummaryDiseaseBreakdownItem = {
  name: string;
  count: number;
};

export interface DashboardSummary {
  totalScans: number;
  healthyCount: number;
  diseasedCount: number;
  invalidCount: number;
  healthScore: number;
  cropsAnalyzed: number;
  lastScanAt?: string | null;
  weeklyTrend: DashboardSummaryWeeklyTrendItem[];
  diseaseBreakdown: DashboardSummaryDiseaseBreakdownItem[];
}

export interface Tip {
  id: string;
  title: string;
  body: string;
  icon: string;
}

export interface Settings {
  language: string;
  /** light | dark | system */
  theme: string;
  /** metric | imperial */
  units: string;
  /** celsius | fahrenheit */
  temperatureUnit: string;
  notificationsEnabled: boolean;
  scanReminders: boolean;
  weatherAlerts: boolean;
  marketingEmails: boolean;
  autoSaveScans: boolean;
  highAccuracyMode: boolean;
  offlineMode: boolean;
}

export interface Profile {
  id: string;
  email: string;
  fullName: string;
  farmName: string;
  location: string;
  bio: string;
  photoUrl: string;
}

export type CreateScanBody = {
  cropId: string;
  /** Data URL of the uploaded leaf image */
  imageDataUrl: string;
};

export type UpdateSettingsBody = {
  language?: string;
  theme?: string;
  units?: string;
  temperatureUnit?: string;
  notificationsEnabled?: boolean;
  scanReminders?: boolean;
  weatherAlerts?: boolean;
  marketingEmails?: boolean;
  autoSaveScans?: boolean;
  highAccuracyMode?: boolean;
  offlineMode?: boolean;
};

export type UpdateProfileBody = {
  fullName?: string;
  farmName?: string;
  location?: string;
  bio?: string;
  photoUrl?: string;
};