# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agro Vision Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List the user's scan history
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: createScan
      tags: [scans]
      summary: Analyze an uploaded crop leaf image
      description: Validates that the uploaded image contains the selected crop's leaf and returns disease analysis. Rejects anything that is not a crop leaf.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cropId: { type: string }
                imageDataUrl:
                  type: string
                  description: Data URL of the uploaded leaf image
              required: [cropId, imageDataUrl]
      responses:
        "200":
          description: Scan analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Invalid image (not a crop leaf, or wrong crop)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanError"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a single scan by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: A scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanError"
    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete a scan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Scan:
      type: object
      properties:
        id: { type: integer }
        cropId: { type: string }
        cropName: { type: string }
        imageDataUrl: { type: string }
        status: { type: string, description: "healthy | diseased | invalid" }
        diseaseName: { type: string, nullable: true }
        confidence: { type: number, nullable: true }
        severity: { type: string, nullable: true }
        summary: { type: string }
        symptoms:
          type: array
          items: { type: string }
        treatment:
          type: array
          items: { type: string }
        prevention:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, cropId, cropName, imageDataUrl, status, summary, symptoms, treatment, prevention, createdAt]
    ScanError:
      type: object
      properties:
        error: { type: string }
        reason: { type: string }
      required: [error, reason]
...